Refactoring into separate packages
Removed some print debugging statements
Added SaveData folder to git-ignore
Further segment code into additional packages: main/parser/ui
