+++ .gitignore
@@ -17,3 +17,4 @@ bin/
 text-ui-test/EXPECTED-UNIX.TXT
 SavedData/SaveData.txt
 SavedData/Save.txt
+SavedData/SaveData.txt
