+++ src/main/java/duke/TaskList.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.events.Task;
 import duke.ui.Ui;
 
 import java.util.List;
