+++ src/main/java/duke/events/Event.java
@@ -1,7 +1,5 @@
 package duke.events;
 
-import duke.Task;
-
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
